Mini Project
It is an application that used Several Button , Entry and as well as  math function.
it has function such as sine,cosine and so on.

Also in the textBasedGame:
You are in a land full of dragons. In front of
you are two caves. In one cave the dragon is friendly will give you his treasure.However a sleeping guard is Front of you and with a surprise. In the other room, the dragon is greedy  with dragon eggs worth more than diamond, Gold and will
eat you on sight